registerpurchase app in the django admin for testing

+++ INVENTORY_MANAGEMENT_SYSTEM/purchase/models.py
@@ -23,7 +23,7 @@ class PurchaseItem(models.Model):
     purchase_order = models.ForeignKey(PurchaseOrder, related_name="items", on_delete=models.CASCADE)
     product = models.ForeignKey(Product, on_delete=models.CASCADE, related_name="purchase_items")
     quantity = models.PositiveIntegerField()
-    cost_price = models.DecimalField(max_digits=10, decimal_places=2)
+    cost_price = models.FloatField()
 
     def __str__(self):
         return f"{self.product.name} x {self.quantity}"
